miscellaneous changes to update more code

+++ src/main/java/me/paulf/wings/client/flight/state/State.java
@@ -46,7 +46,7 @@ public abstract class State {
     }
 
     private float getPitch(double x, double y, double z) {
-        return Mth.toDegrees((float) -Math.atan2(y, net.minecraft.util.Mth.sqrt(x * x + z * z)));
+        return Mth.toDegrees((float) -Math.atan2(y, net.minecraft.util.Mth.sqrt((float) (x * x + z * z))));
     }
 
     public final void beginAnimation(Animator animator) {
